cleanup(vision-v1) pass 1: dead code, dedup, stale refs, resolver split
Audit-driven cleanup of GeecsBluesky + GeecsCAGateway (findings and
keep/defer decisions recorded in Planning/cleanup_vision_v1/00_overview.md).
No behavior change; both suites green (407 + 158).

GeecsBluesky 0.27.0:
- extract ConfigResolver/ConfigsRepoResolver -> config_resolver.py
  (re-exported from scan_request_runner; import surface unchanged)
- delete dead: dialog_queue/restore_failures/_ScanConfig/_CONNECT_TIMEOUT
  (bridge), singular resolve_save_set_checked chain (superseded by the
  plural M4 forms; tests ported), utils.build_signal_attrs (never wired in)
- dedup: bridge _build_positions delegates to session._positions
- fix ~20 stale docstring refs to deleted Geecs* direct-backend classes;
  rewrite step_scan example off the deleted host/port constructor

GeecsCAGateway 0.12.0:
- delete dead: GeecsDb.list_devices, FakeGeecsDevice.fire_shot
- dedup: shared transport/_coerce.coerce_scalar (tcp_subscriber + udp_client)
- fix stale config.py/naming.py docstrings and CLAUDE.md fake-server claim
