+++ GeecsBluesky/tests/test_utils.py
@@ -1,10 +1,10 @@
-"""Tests for safe_name / build_signal_attrs and device legacy header maps."""
+"""Tests for safe_name and device legacy header maps."""
 
 from __future__ import annotations
 
 import pytest
 
-from geecs_bluesky.utils import build_signal_attrs, safe_name
+from geecs_bluesky.utils import safe_name
 
 pytest.importorskip("aioca")  # devices are CA-backed
 
@@ -21,15 +21,6 @@ def test_safe_name_mangles_and_lowercases() -> None:
     assert safe_name("Wavelength (nm)") == "wavelength__nm"
     assert safe_name("Position.Axis 1") == "position_axis_1"
     assert safe_name("") == "var"
-
-
-def test_build_signal_attrs_disambiguates_collisions() -> None:
-    # Two variables that mangle to the same attr get _2 appended in order.
-    pairs = build_signal_attrs(["A (x)", "A [x]", "B"])
-    attrs = [attr for attr, _ in pairs]
-    assert attrs == ["a__x", "a__x_2", "b"]
-    # The original variable name is preserved alongside each attr.
-    assert pairs == [("a__x", "A (x)"), ("a__x_2", "A [x]"), ("b", "B")]
 
 
 def test_generic_detector_column_headers() -> None:
